+++ src/server.jsx
@@ -10,6 +10,7 @@ import { keys } from 'lodash';
 import Raven from 'raven';
 import cookie, { plugToRequest } from 'react-cookie';
 import locale from 'locale';
+import { detect } from 'detect-browser';
 
 import routes from '~/routes';
 import nlLocale from '~/../locales/nl.json';
@@ -62,6 +63,8 @@ server
     const url = req.originalUrl || req.url;
     const location = parseUrl(url);
 
+    const browserdetect = detect(req.headers['user-agent']);
+
     const lang = new locale.Locales(
       cookie.load('lang') || req.headers['accept-language'],
     )
@@ -78,6 +81,7 @@ server
         locale: lang,
         messages: locales[lang],
       },
+      browserdetect,
     };
     const history = createMemoryHistory({
       initialEntries: [req.url],