+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changes
 
+- Browser detect feature, adding a deprecation message for ancient browsers in the `App` component @sneridagh
+
 ## 4.0.0-alpha.35 (2020-01-31)
 
 ### Changes